Refactor script transaction submission into separate module
The three server scripts share much transaction submission functionality.
This refactor puts the functionality in a single DRY module, and will
simplify adding additional functionality.
